backend/requirements.txt
Django>=4.2
django-cors-headers
djangorestframework
psycopg2-binary
